remove tron bridge
