+++ libs/coin-modules/coin-tron/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./types";
 
-export { createBridges } from "./bridge/index";
-export { isAccountEmpty } from "./bridge/utils";
+export { isAccountEmpty } from "./resources";
