+++ libs/ledger-live-common/src/transaction/deviceTransactionConfig.ts
@@ -7,7 +7,7 @@ import type { ExtraDeviceTransactionField as ExtraDeviceTransactionField_casper 
 import type { ExtraDeviceTransactionField as ExtraDeviceTransactionField_filecoin } from "@ledgerhq/coin-filecoin/bridge/deviceTransactionConfig";
 import type { ExtraDeviceTransactionField as ExtraDeviceTransactionField_stacks } from "@ledgerhq/coin-stacks/bridge/deviceTransactionConfig";
 import type { ExtraDeviceTransactionField as ExtraDeviceTransactionField_polkadot } from "@ledgerhq/coin-polkadot/bridge/deviceTransactionConfig";
-import type { ExtraDeviceTransactionField as ExtraDeviceTransactionField_tron } from "@ledgerhq/coin-tron/bridge/deviceTransactionConfig";
+import type { ExtraDeviceTransactionField as ExtraDeviceTransactionField_tron } from "@ledgerhq/coin-tron/deviceTransactionConfig";
 
 type ExtraDeviceTransactionField =
   | ExtraDeviceTransactionField_casper
